feat: check methods - routes
