+++ config/basic.conf
@@ -1,8 +1,9 @@
 server {
-	listen = 1030
+	listen = 1031
 	route /cgi {
 		root = /content/cgi
 		index = current-time.py
+		allow_methods = GET POST DELETE
 	}
 	route /redirect {
 		redirect = https://profile.intra.42.fr/
@@ -12,8 +13,10 @@ server {
 server {
 	listen = 1027
 	server_name = server 1027
+
 	route /css {
 		root = /content/css
+		allow_methods = POST
 		autoindex = true
 	}
 }
