+++ content/page-errors/500.html
@@ -2,7 +2,7 @@
 <html lang="en" >
 <head>
   <meta charset="UTF-8">
-  <title>CodePen - 404 error message</title>
+  <title>CodePen - 500 error message</title>
   <link rel="stylesheet" href="./css/error.css">
 
 </head>
@@ -12,10 +12,10 @@
 <link href="https://fonts.googleapis.com/css?family=Poppins" rel="stylesheet">
 </head>
 <body>
-  <div class="message">Page not found.</div>
-  <div class="message2">You tried to access a page that does not exist or has been removed.</div>
+  <div class="message">Internal Server Error.</div>
+  <div class="message2">The server encountered an internal error and was unable to complete your request.</div>
   <div class="container">
-    <div class="neon">404</div>
+    <div class="neon">500</div>
     <div class="door-frame">
       <div class="door">
         <div class="rectangle">
