+++ src/class/Client/CGI.cpp
@@ -1,28 +1,28 @@
-#include "Response.hpp"
+#include "Client.hpp"
 
 extern char **environ;
 
-void Response::freeEnviron(char **envp)
+void Client::freeEnviron(char **envp)
 {
     for (int i = 0; envp[i]; i++)
         delete[] envp[i];
     delete[] envp;
 }
 
-bool Response::checkFile(const string &file)
+bool Client::checkFile(const string &file)
 {
     ifstream f(file.c_str());
     return (f.good());
 }
 
-char **Response::configEnviron(Server& server, Request &req)
+char **Client::configEnviron(Server& server, Request &req)
 {
     char **envp = new char *[7];
     
     string serverName = "SERVER_NAME=" + server.getServerName();
     string serverPort = "SERVER_PORT=" + Utils::itostr(server.getListen());
-    string protocol = "SERVER_PROTOCOL=" + req._version;
-    string pathInfo = "PATH_INFO=" + req._uri;
+    string protocol = "SERVER_PROTOCOL=" + _request.getVersion();
+    string pathInfo = "PATH_INFO=" + _request.getURI();
     string method = "REQUEST_METHOD=" + req.getMethod();
     string gatewayInterface = "GATEWAY_INTERFACE=CGI/1.1";
 
@@ -36,14 +36,14 @@ char **Response::configEnviron(Server& server, Request &req)
     return (envp);
 }
 
-string Response::executeCGI(Request &req, Server& server, string filePath)
+string Client::executeCGI(Request &req, Server& server, string filePath)
 {
     int pid;
     int fd[2];
     int status;
     string result;
     if (!checkFile(filePath)) {
-        _statusCode = NOT_FOUND;
+        _response.setStatusCode(NOT_FOUND);
         return ("");
     }
 
@@ -59,11 +59,11 @@ string Response::executeCGI(Request &req, Server& server, string filePath)
         throw Server::exception(RED "Error: Fork failed" RESET);
     }
     if (pid == 0) {
-        signal(SIGINT, SIG_IGN);
-        signal(SIGTERM, SIG_IGN);
+        signal(SIGINT, Utils::handleSignals);
+        signal(SIGTERM, Utils::handleSignals);
         char **envp = configEnviron(server, req);
         char *args[] = {const_cast<char *>(_executor.c_str()), const_cast<char *>(filePath.c_str()), NULL};
-
+    
         dup2(fd[1], STDOUT_FILENO);
         close(fd[0]);
         if (execve(_executor.c_str(), args, envp) == -1) {
@@ -87,13 +87,13 @@ string Response::executeCGI(Request &req, Server& server, string filePath)
     return (result);
 }
 
-string Response::readCGI(int fd_in)
+string Client::readCGI(int fd_in)
 {
     ssize_t bytesRead;
     char buffer[65535];
     string result;
 
-    while ((bytesRead = read(fd_in, &buffer, sizeof(buffer))) > 0) {
+    while ((bytesRead = read(fd_in, &buffer, sizeof(buffer) - 1)) > 0) {
         buffer[bytesRead] = '\0';
         result += buffer;
     }
