+++ src/class/Run/Run.cpp
@@ -149,89 +149,3 @@ void Run::startServer(vector<Server>& servers)
         }
     }
 }
-
-// void Run::startServer(vector<Server>& servers){
-//     Client client;
-//     int returnValue;
-//     vector<struct pollfd> pollFds = loadPolls(servers);
-//     vector<Client *> clients;
-
-
-//     while (true){
-//         returnValue = poll(pollFds.data(), pollFds.size(), 60 * 1000);
-//         if (signalUsed){
-//             break ;
-//         }
-//         else if (returnValue == -1){
-//             break ;
-//         }
-
-//         else{
-//             for (size_t i = 0; i < pollFds.size(); i++){
-//                 if (pollFds[i].revents & POLLIN && i < servers.size()){
-//                     int clientFd = acceptNewConnection(pollFds[i].fd, pollFds);
-//                     client.setServer(servers[i]);
-//                     _mapClient[clientFd] = &client;
-//                 }
-//                 else if (pollFds[i].revents & POLLIN){
-//                     Client *actual = getClientFd(pollFds[i].fd);
-//                     actual->getRequest()->readRequest(pollFds, i);
-//                 }
-//                 else if (pollFds[i].revents & POLLOUT){ // response
-//                     Client *response = getClientFd(pollFds[i].fd);
-//                     response->sendResponse(pollFds[i].fd);
-//                     _mapClient.erase(pollFds[i].fd);
-//                     pollFds.erase(pollFds.begin() + i);
-//                 }
-//             }
-//         }
-//     }
-// }
-
-// void Run::startServer(vector<Server>& servers)
-// {
-//     int pollValue;
-//     while (true) {
-//         Client client;
-//         bool requestFound = false;
-//         size_t i = 0;
-
-//         if (signalUsed){
-//             break ;
-//         }
-//         while (i < servers.size()) {
-//             pollValue = poll(&servers[i].getPollFd(), 1, 10);
-//             if (signalUsed){
-//                 break ;
-//             }
-//             if (pollValue == -1)
-//                 throw Server::exception(RED "Error: poll failed" RESET);
-//             if (servers[i].getPollFd().revents & POLLIN) {
-//                 try {
-//                         requestFound = true;
-//                         struct pollfd actualFd = Run::acceptNewConnection(servers[i].getPollFd().fd);
-//                         servers[i].setClientFd(actualFd);
-//                         client.setServer(servers[i]);
-//                     while (client.getRequest()->getIsReadyForResponse() == false) {
-//                         cout << "getting request" << endl;
-//                         client.getRequest()->readRequest(actualFd);
-//                     }
-//                 }
-//                 catch (const std::exception& e) {
-//                     cerr << "Error reading request: " << e.what() << endl;
-//                 }
-//             }
-//             else if (servers[i].getPollFd().revents & POLLOUT) {
-//                 try {
-//                     requestFound = false;
-//                     client.sendResponse();
-//                     servers[i].getServerFd();
-//                 }
-//                 catch (const std::exception& e) {
-//                     cerr << "Error sending response: " << e.what() << endl;
-//                 }
-//             }
-//             i = requestFound ? i : i + 1;
-//         }
-//     }
-// }
